refactor(examples-advanced): renamed spawning "crate" to blueprints & cleaned up  * renamed spawning to blueprints  * moved most code out of mod to spawn_from_blueprints  * renamed a few modules  * moved a few components around for clarity & coherence
