+++ examples/advanced/core/blueprints/clone_entity.rs
