+++ examples/advanced/game/in_game.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 
-use crate::{assets::GameAssets, core::spawning::GameWorldTag, state::{InAppRunning, GameState}};
+use crate::{assets::GameAssets, core::blueprints::GameWorldTag, state::{InAppRunning, GameState}};
 
 pub fn setup_game(
     mut commands: Commands,
